Reading Level Size from XML
Reading Player Spawn Point from XML

+++ SpaceMarine/src/sm_entity.cpp
@@ -47,8 +47,8 @@ void SMEntity::SetID(const int iID)
 bool SMEntity::VOnLoad(const Base::cString & FilePath, const int Width,
 	const int Height, const int MaxFrames)
 {
-	m_Size.x = Width;
-	m_Size.y = Height;
+	m_Size.x = static_cast<float>(Width);
+	m_Size.y = static_cast<float>(Height);
 
 	m_pSurface = SMSurface::OnLoad(FilePath);
 	if (m_pSurface == NULL)
